module_paper_trading/requirement.txt
google-cloud-pubsub 
google-cloud-firestore
